Medex con Frecuencia de Mantenimiento Preventivo

+++ app/Http/Controllers/EquiposCtrl.php
@@ -11,6 +11,7 @@ namespace App\Http\Controllers;
     //end modelos
 
 use Zxing\QrReader;
+use Illuminate\Support\Collection;
 use Illuminate\Http\Request;
 use SimpleSoftwareIO\QrCode\Facades\QrCode;
 
@@ -23,6 +24,7 @@ class EquiposCtrl extends Controller
      */
     public function index(Request $request)
     {
+        $Equipos=equipo::all();
         if(! $request->user()){
             //Toastr::warning('Es necesario iniciar sesión para validar el acceso a los datos', 'Acceso denegado');
             return view('auth.login');
@@ -34,14 +36,13 @@ class EquiposCtrl extends Controller
 
     public function qr(Request $request)
     {    
-        $Equipo = Equipo::where('ID_inventario', $request->input('valor'))->get(
-
-        );
-        dd($Equipo);
-        $NomEquipo = $Equipo->Nombre;
-        $NumSerie = $Equipo->Num_de_serie;
-        $QRcode = QrCode::size(500)->generate($request->input('valor'));
-        $pdfDoc = \PDF::loadView('Formatos.QR',["valor" => $QRcode]);
+        $Equipo = Equipo::where('ID_inventario', $request->input('valor'))->get();
+        $QRvalor = $Equipo[0]['ID_inventario'];
+        $NomEquipo = $Equipo[0]['Nombre'];
+        $id = $Equipo[0]['ID_inventario'];
+        $NumSerie = $Equipo[0]['Num_de_serie'];
+        $QREquipo = QrCode::size($request->input('Tamano'))->generate('http://protomedex.site/Equipos/'.$QRvalor);
+        $pdfDoc = \PDF::loadView('Formatos.QR',["valor" => $QREquipo],compact(['NumSerie', 'NomEquipo', 'id']));
         return $pdfDoc->stream($NomEquipo.'-'.$NumSerie.'.pdf');
     }
 
@@ -58,10 +59,14 @@ class EquiposCtrl extends Controller
      */
     public function create(Request $request)
     {
-        //$request->user()->authorizeRole(['superAdmin', 'Ingeniero']);
         $Marcas = Marca::all();
         $Mntos = Mnto::all();
-         return view('Equipo.AddToInventario', compact('Mntos'), compact('Marcas'));
+        if(! $request->user()){
+            return view('auth.login');
+        } elseif($request->user()->authorizeRole(['superAdmin', 'Ingeniero'])) {
+            return view('Equipo.AddToInventario', compact('Mntos'), compact('Marcas'));
+        }
+        //$request->user()->authorizeRole(['superAdmin', 'Ingeniero']);
     }
 
     /**
@@ -72,17 +77,44 @@ class EquiposCtrl extends Controller
      */
     public function store(Request $request)
     {
-      
+        /**$Req       = $request->input('requerimientosEq');
+        $Riesgo    = $request->input('riesgoEq');
+        $Funcion   = $request->input('funcionEq');
+        $Contrib   = $request->input('contribucionTratamiento');
+        $Frec      = $request->input('frecUso');
+        $Confiabl  = $request->input('confiabilidad');
+        $Facilidad = $request->input('facilidadUso');
+
+        $Ec = (($Frec + $Confiabl + $Facilidad + $Contrib)*10)/28;
+        $FMP = $Ec + $Req + $Riesgo + $Funcion;
+$request->input('requerimientosEq riesgoEq funcionEq');
+        
+        dd($dd);*/
+
         if($request->hasFile('imagenEquipo')){
             $file = $request->file('imagenEquipo');
             $name = time().$file->getClientOriginalName();
             $file->move(public_path().'/imgInv/', $name);
         }
+        $Ec = 0;
+        $Req       = $request->input('requerimientosEq');
+        $Riesgo    = $request->input('riesgoEq');
+        $Funcion   = $request->input('funcionEq');
+        $FMP       = $Req + $Riesgo + $Funcion + $Ec;
+            switch ($FMP) {
+                case $FMP<=25 && $FMP>12:
+                    $dd= "I";
+                    break;
+                case $FMP<=28 && $FMP>25:
+                    $dd= "II";
+                    break;
+                case $FMP<=30 && $FMP>28:
+                    $dd= "III";
+                    break;
+            }
 
-       
-               
                      $AddEquipo = new Equipo();
-        $IdEquipo  = $AddEquipo->ID_inventario = $request->input('ID_inventario');
+        $IdEquipo  = $AddEquipo->ID_inventario = $Funcion.$Riesgo.$Req.time();
         $NomEquipo = $AddEquipo->Nombre = $request->input('Nombre');
                      $AddEquipo->Area = $request->input('Area');
                      $AddEquipo->Tipo = $request->input('Tipo');
@@ -94,11 +126,11 @@ class EquiposCtrl extends Controller
                      $AddEquipo->vencimientoGarantia = $request->input('vencimientoGarantia');
                      $AddEquipo->Consumo_electrico = $request->input('Consumo_electrico');
                      $AddEquipo->Mnto = $request->input('Mnto');
-                     //$AddEquipo->porcentUso = $request->input('porcentUso');
                      $AddEquipo->ultimoMantenimiento = $request->input('ultimoMantenimiento');
-                     $AddEquipo->prioridad = $request->input('prioridad');//gregarlo como boton en el frm y como campo en elmigration
-                     $AddEquipo->imagenEquipo = $name;
+                     $AddEquipo->prioridad = $dd;
+                     $AddEquipo->imagenEquipo = $name; 
                      $AddEquipo->save();
+                     
 
 
         $Equipos = Equipo::all();
@@ -121,7 +153,7 @@ class EquiposCtrl extends Controller
     public function filtrarInventario(Request $request)
     {
         $valor = $request->valor;
-        $Equipos = Equipo::where('Nombre','LIKE',$request->valor)->orWhere('Modelo','LIKE',$request->valor)->orWhere('ID_inventario','LIKE',$request->valor)->get();
+        $Equipos = Equipo::where('Nombre','LIKE',$request->valor)->orWhere('Modelo','LIKE',$request->valor)->orWhere('ID_inventario','LIKE',$request->valor)->orWhere('Estatus','LIKE',$request->valor)->orWhere('Mnto','LIKE',$request->valor)->orWhere('Marca','LIKE',$request->valor)->orWhere('Area','LIKE',$request->valor)->get();
         return view('Equipo.Inventario',compact('Equipos'));
     }
 
@@ -133,44 +165,40 @@ class EquiposCtrl extends Controller
      */
     public function show(Request $request, $ID_inventario)
     {
-        //$request->user()->authorizeRole(['superAdmin', 'Ingeniero', 'personalMedico']);
         $Equipo = Equipo::find($ID_inventario);
-        if($Equipo->prioridad == 'II'){             
+        if(! $request->user()){
+            //Toastr::warning('Es necesario iniciar sesión para validar el acceso a los datos', 'Acceso denegado');
+            return view('auth.login');
+        } elseif($request->user()->authorizeRole(['superAdmin', 'Ingeniero'])) {
 
-                    $fecha_act = $Equipo->ultimoMantenimiento; 
-            
-                    $fechas[0]= date("d-m-Y",strtotime($fecha_act."+ 4 month")); 
-                    $fechas[1]= date("d-m-Y",strtotime($fechas[0]."+ 4 month"));
-                    $fechas[2]= date("d-m-Y",strtotime($fechas[1]."+ 4 month"));
-                    $fechas[3]= date("d-m-Y",strtotime($fechas[2]."+ 4 month"));
-                    $fechas[4]= date("d-m-Y",strtotime($fechas[3]."+ 4 month"));
-                    $fechas[5]= date("d-m-Y",strtotime($fechas[4]."+ 4 month"));
-        } elseif ($Equipo->prioridad == 'III'){
+            if($Equipo->prioridad == 'II'){             
 
                 $fecha_act = $Equipo->ultimoMantenimiento; 
+        
+                $fechas[0]= date("d-m-Y",strtotime($fecha_act."+ 4 month")); 
+                $fechas[1]= date("d-m-Y",strtotime($fechas[0]."+ 4 month"));
+                $fechas[2]= date("d-m-Y",strtotime($fechas[1]."+ 4 month"));
+                //$fechas[3]= date("d-m-Y",strtotime($fechas[2]."+ 4 month"));
+    } elseif ($Equipo->prioridad == 'III'){
 
-                $fechas[0]= date("d-m-Y",strtotime($fecha_act."+ 6 month")); 
-                $fechas[1]= date("d-m-Y",strtotime($fechas[0]."+ 6 month"));
-                $fechas[2]= date("d-m-Y",strtotime($fechas[1]."+ 6 month"));
-                $fechas[3]= date("d-m-Y",strtotime($fechas[2]."+ 6 month"));
-        }elseif ($Equipo->prioridad == 'I'){
+            $fecha_act = $Equipo->ultimoMantenimiento; 
 
-                $fecha_act = $Equipo->ultimoMantenimiento; 
+            $fechas[0]= date("d-m-Y",strtotime($fecha_act."+ 6 month")); 
+            $fechas[1]= date("d-m-Y",strtotime($fechas[0]."+ 6 month"));
+            $fechas[2]= date("d-m-Y",strtotime($fechas[1]."+ 6 month"));
+            //$fechas[3]= date("d-m-Y",strtotime($fechas[2]."+ 6 month"));
+    }elseif ($Equipo->prioridad == 'I'){
 
-                $fechas[0]= date("d-m-Y",strtotime($fecha_act."+ 2 month")); 
-                $fechas[1]= date("d-m-Y",strtotime($fechas[0]."+ 2 month"));
-                $fechas[2]= date("d-m-Y",strtotime($fechas[1]."+ 2 month"));
-                $fechas[3]= date("d-m-Y",strtotime($fechas[2]."+ 2 month"));
-                $fechas[4]= date("d-m-Y",strtotime($fechas[3]."+ 2 month")); 
-                $fechas[5]= date("d-m-Y",strtotime($fechas[4]."+ 2 month"));
-                $fechas[6]= date("d-m-Y",strtotime($fechas[5]."+ 2 month"));
-                $fechas[7]= date("d-m-Y",strtotime($fechas[6]."+ 2 month"));
-                $fechas[8]= date("d-m-Y",strtotime($fechas[7]."+ 2 month")); 
-                $fechas[9]= date("d-m-Y",strtotime($fechas[8]."+ 2 month"));
-                $fechas[10]= date("d-m-Y",strtotime($fechas[9]."+ 2 month"));
-                $fechas[11]= date("d-m-Y",strtotime($fechas[10]."+ 2 month"));
+            $fecha_act = $Equipo->ultimoMantenimiento; 
+
+            $fechas[0]= date("d-m-Y",strtotime($fecha_act."+ 12 month")); 
+            $fechas[1]= date("d-m-Y",strtotime($fechas[0]."+ 12 month"));
+    }
+            return view('Equipo.ShowEquipo', compact('Equipo'),  compact('fechas'));
         }
-        return view('Equipo.ShowEquipo', compact('Equipo'),  compact('fechas'));
+
+        
+
     }
 
     /**
@@ -181,12 +209,33 @@ class EquiposCtrl extends Controller
      */
     public function edit(Request $request, $ID_inventario)
     {
-        //$request->user()->authorizeRole(['superAdmin', 'Ingeniero']);
         $Marcas = Marca::all();
         $Mntos = Mnto::all();
-        $Equipo = Equipo::find($ID_inventario);
-        //dd($Marcas);
-        return view('Equipo.editEquipo', compact(['Marcas', 'Mntos', 'Equipo']));
+        $Equipo = Equipo::find($ID_inventario);  
+        
+        $Ec = 0;
+        $Req       = $request->input('requerimientosEq');
+        $Riesgo    = $request->input('riesgoEq');
+        $Funcion   = $request->input('funcionEq');
+        $FMP       = $Req + $Riesgo + $Funcion + $Ec;
+            switch ($FMP) {
+                case $FMP<=25 && $FMP>12:
+                    $dd= "I";
+                    break;
+                case $FMP<=28 && $FMP>25:
+                    $dd= "II";
+                    break;
+                case $FMP<=30 && $FMP>28:
+                    $dd= "III";
+                    break;
+            }
+
+
+        if(! $request->user()){
+            return view('auth.login');
+        } elseif($request->user()->authorizeRole(['superAdmin', 'Ingeniero'])) {
+            return view('Equipo.editEquipo', compact(['Marcas', 'Mntos', 'Equipo']));
+        }
     }
 
     /**
@@ -219,10 +268,16 @@ class EquiposCtrl extends Controller
     public function destroy($ID_inventario)
     {
         $Equipo = Equipo::find($ID_inventario);
+        if(! $request->user()){
+            //Toastr::warning('Es necesario iniciar sesión para validar el acceso a los datos', 'Acceso denegado');
+            return view('auth.login');
+        } elseif($request->user()->authorizeRole(['superAdmin', 'Ingeniero', 'personalMedico'])) {
+            $Equipo->delete();
+            return view('Nav.Navegacion');
+        } 
+        
         //dd($Equipo);
-        $Equipo->delete();
-        $Equipo = Equipo::all();
-        return view('Nav.Navegacion');
+
     }
 
     public function FrecuenciaMP(Request $request)
